// frontend/src/generated/workspace-api/types.gen.ts
export type PhaseDocumentCreate = {
    /**
     * Title
     */
    title: string;
    /**
     * Content
     */
    content?: string | null;
    /**
     * Document Type
     */
    document_type?: string;
};

/**
 * PhaseDocumentUpdate
 */
export type PhaseDocumentUpdate = {
    /**
     * Title
     */
    title?: string | null;
    /**
     * Content
     */
    content?: string | null;
    /**
     * Document Type
     */
    document_type?: string | null;
};

/**
 * PhaseUpdate
 */
export type PhaseUpdate = {
    /**
     * Status
     */
    status?: string | null;
    /**
     * Start Date
     */
    start_date?: string | null;
    /**
     * End Date
     */
    end_date?: string | null;
    /**
     * Notes
     */
    notes?: string | null;
};